pyproject.toml
[project]
name = "facelift"
version = "0.1.0"
description = "FaceLift: Single Image 3D Face Reconstruction"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "torch>=2.4.0",
    "torchvision>=0.19.0",
    "transformers>=4.44.0",
    "diffusers[torch]>=0.30.0",
    "huggingface-hub>=0.35.0",
    "xformers>=0.0.27",
    "accelerate>=0.33.0",
    "Pillow>=10.4.0",
    "opencv-python>=4.10.0",
    "scikit-image>=0.21.0",
    "lpips>=0.1.4",
    "numpy>=1.26.0",
    "matplotlib>=3.7.0",
    "scikit-learn>=1.3.0",
    "einops>=0.8.0",
    "jaxtyping>=0.2.19",
    "pytorch-msssim>=1.0.0",
    "easydict>=1.13",
    "pyyaml>=6.0.0",
    "wandb>=0.19.0",
    "termcolor>=2.4.0",
    "plyfile>=1.0.3",
    "tqdm",
    "gradio>=5.49.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-multipart>=0.0.17",
    "requests>=2.32.0",
    "facenet-pytorch",
    "rembg",
    "ffmpeg-python>=0.2.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.8.0",
    "pre-commit>=4.0.0",
]
raster = [
    "git+https://github.com/graphdeco-inria/diff-gaussian-rasterization",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.ruff]
target-version = "py310"
